Added the code for damaging stuff
Added a destructable object that reduces its health after it gets hit. Added a function in that object that checks its hp each tick. Added code to the player character that allows it to do damage. Switched the hit detection to detecting static objects so it could detect the new temporary damageable object.