Cleaned up code, boss has multiple attacks
Worked on making less code to the same thing and parenting blueprints for the boss.
Attacks can get cooldown and animations for attacks can be listed in the boss blueprint using the details tab.
The cooldown values on those attacks are currently done using the behavior tree instead of the code in the blueprint.